[Mod] #1 Entity에 ERD 내용 반영 - Table, UniqueConstraints 추가

+++ src/main/java/com/nova/yeobaek/domain/item/domain/Category.java
@@ -43,7 +43,7 @@ public class Category extends BaseEntity {
 	private String name;
 
 	@Column(nullable = false)
-	private int level;
+	private Long level;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "parent_id")
